requirements.txt
boto3==1.26.86
minio==7.1.13
psycopg2-binary==2.9.5
jsonschema==3.2.0